Fix float conversion

+++ framework/foundation/include/math/math_helpers.h
@@ -48,12 +48,12 @@ static inline f32 f32v3_dot(f32v3 a, f32v3 b)
 
 static inline f32 angle_between_points(f32v2 a, f32v2 b)
 {
-	return atan2(a.y - b.y, a.x - b.x);
+	return atan2f(a.y - b.y, a.x - b.x);
 }
 
 static inline f32 rad_to_deg(f32 rad)
 {
-	return rad / (M_PI * 2.f) * 360.f;
+	return rad / ((f32)M_PI * 2.f) * 360.f;
 }
 
 static inline f32v2 tile_position_to_position(i32v2 tile_position, i32v2 tile_size)
